requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-openai>=0.0.2
langchain-huggingface>=0.0.2
sentence-transformers>=2.2.2
numpy>=1.23.0
faiss-cpu>=1.7.4
sqlparse>=0.4.4
python-dotenv>=1.0.0
pydantic>=2.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.23.0
sqlite3-api>=2.0.0
pyyaml>=6.0.0

# Database connectors
psycopg2-binary>=2.9.9  # PostgreSQL support
neo4j>=5.15.0  # Neo4j support

# Optional dependencies
pyarrow>=15.0.0  # For efficient data handling in Neo4j
pytest>=7.4.0  # For testing
pytest-asyncio>=0.21.0  # For async tests 